+++ src/App.tsx
@@ -1,14 +1,13 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import { useEffect, useCallback, useState } from 'react';
-import cn from 'classnames';
+import { useEffect, useState } from 'react';
 import { useTodoContext } from './context/TodoContext';
-import { getTodos, updateTodo } from './api/todos';
+import { getTodos } from './api/todos';
 import { Error } from './types/Error';
 import { TodoList } from './components/TodoList';
-import { TodoForm } from './components/TodoForm';
 import { FooterFilter } from './components/FooterFilter';
 import { ErrorMessage } from './components/ErrorMessage';
 import { Todo } from './types/Todo';
+import { Header } from './components/Header';
 
 const USER_ID = 10407;
 
@@ -18,14 +17,11 @@ export const App: React.FC = () => {
     error,
     setTodos,
     setError,
-    setTodoIdsInUpdating,
   } = useTodoContext();
 
-  const [toggleStatus, setToggleStatus] = useState<boolean>(false);
   const [completedTodos, setCompletedTodos] = useState<Todo[]>([]);
 
   useEffect(() => {
-    setToggleStatus(todos.every(todo => todo.completed));
     setCompletedTodos(todos?.filter(todo => !todo.completed));
   }, [todos]);
 
@@ -40,38 +36,6 @@ export const App: React.FC = () => {
     }
   };
 
-  const toggleAllComplete = useCallback(async (): Promise<void> => {
-    if (todos.every(t => t.completed) || todos.every(t => !t.completed)) {
-      setTodoIdsInUpdating(todos.map(todo => todo.id));
-    } else {
-      setTodoIdsInUpdating(
-        todos.filter(todo => !todo.completed).map(todo => todo.id),
-      );
-    }
-
-    try {
-      await Promise.all(todos.map(todo => {
-        const updatedTodo = {
-          ...todo,
-          completed: !toggleStatus,
-        };
-
-        return updateTodo(todo.id, updatedTodo);
-      }));
-
-      setTodos(prevTodos => prevTodos.map(todo => ({
-        ...todo,
-        completed: !toggleStatus,
-      })));
-
-      setToggleStatus(!toggleStatus);
-    } catch {
-      setError(Error.UPDATE);
-    } finally {
-      setTodoIdsInUpdating([]);
-    }
-  }, [todos, toggleStatus]);
-
   useEffect(() => {
     loadTodos();
   }, []);
@@ -81,17 +45,7 @@ export const App: React.FC = () => {
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          <button
-            type="button"
-            className={cn('todoapp__toggle-all', {
-              active: todos.every((todo) => todo.completed),
-            })}
-            onClick={toggleAllComplete}
-          />
-
-          <TodoForm />
-        </header>
+        <Header />
 
         {todos.length > 0 && (
           <>
